+++ packages/web/src/components/play-bar/desktop/components/SocialActions.tsx
@@ -15,7 +15,7 @@ import FavoriteButton from 'components/alt-button/FavoriteButton'
 import RepostButton from 'components/alt-button/RepostButton'
 import Tooltip from 'components/tooltip/Tooltip'
 import { GatedConditionsPill } from 'components/track/GatedConditionsPill'
-import { useAuthenticatedClickCallback } from 'hooks/useAuthenticatedCallback'
+import { useRequiresAccountOnClick } from 'hooks/useRequiresAccount'
 import { shouldShowDark } from 'utils/theme/theme'
 
 import styles from './SocialActions.module.css'
@@ -60,7 +60,7 @@ export const SocialActions = ({
   const gatedTrackStatus = trackId && gatedTrackStatusMap[trackId]
   const { onOpen: openPremiumContentPurchaseModal } =
     usePremiumContentPurchaseModal()
-  const onClickPill = useAuthenticatedClickCallback(() => {
+  const onClickPill = useRequiresAccountOnClick(() => {
     openPremiumContentPurchaseModal(
       { contentId: trackId, contentType: PurchaseableContentType.TRACK },
       { source: ModalSource.PlayBar }
